Always use redacted icon for the shelf
When the notif is sensitive, we were setting the redacted icon in the
shelf in createIcons, but not in updateIcons. This could have caused an
issue where the non-redacted icon would be visible in the shelf on the
lockscreen in spite of redaction settings.

Also formatted the file with ktfmt.

Test: IconManagerTest (added an extra check) + post some priority
conversations and check the icons on lockscreen shelf and aod
Flag: NONE
Fix: 326094524

Change-Id: I001d4b745792648386301810f85effac9ef026c8