Fix IME nav bar long click accessibility toggle
The buttons drawn in the IME nav bar (IME collapse and IME switcher) are
incorrectly set as longClickable (as View#setOnLongClickListener sets
this property to true even when the passed in argument is null).

This lead to Switch Access for accesibility showing a menu with both
Select and Touch & Hold options, but the latter would have no effect.
Instead, we now correctly set these buttons as not longClickable, which
omits showing the menu, and executes the Select action directly.

Test: n/a
Bug: 315723734
Change-Id: I4241d8dee0bea238cd73161a7d8713a03db48ee8