Set sessionId in autofillIds in CredentialAutofillService
There was a bug displaying autofill suggestion chips for credman. The
bug was due to mismatching autofillId hash between the autofill session and
the autofill ids stored in credential option. The autofill ids stored in
credential option were created before the autofill session processed
them, and do not have the session id associated with them. Because of
this, when the autofill session was processing the response from
credential autofill service, it failed to update the viewState with the
correct autofill id hash.

Bug: 327319999
Test: addressbook
Change-Id: I5abba6ee86936d1feeee1904b781146eb211a9a7