Do not create indicator before startDragToDesktop.
If we are not past the transition height used in
DesktopModeVisualIndicator and the MoveToDesktopAnimator has not been
created, do not create a visual indicator. This avoids an issue with
split tasks where dragging above the status bar on the focused task
would show an indicator over the unfocused task while still in split
state.

Once the animator has started, indicator works like normal.

Bug: 328423886
Test: Manual
Change-Id: I35c505ceceb70b410526f97b0b609f0c4279dec3