Create Flow States, UI, BiometricFlow, and Parsing
Similar to the Get flow, this combines 4 previously scoped out bugs.

This adds on all the state management for the single tap create flow,
and sets up detailed conditionals.

This flow can *already* be triggered (as can the get flow) in the present setup,
and further chains will complete the feature. Thus, we introduce an E2E
with which we can continue extending. For example, there have been
recent alignments in UX with the BiometricTeam, and those iterations can
follow once the E2E is embedded, especially since we will continue to
utilize parameters that are in the process of being released.

Bug: 327619148
Bug: 327620245
Bug: 327620327
Bug: 327621520
Test: Junit tests, UX tests, and build tests (flow not triggerable in
normal use cases yet)

Change-Id: I5532662315e9a06b6a74f4454418ee1ff3cd243d