Don't stop or restart hub timeout when dreaming state changes
While we only want to actually close the hub if the timeout is
triggered while the user is dreaming, we don't want starting or
stopping dreaming to affect the timer as it isn't a user interaction
and happens underneath the hub.

This means that if the screen timeout is 2 minutes, opening the hub
over the lock screen, spending 1 minute idle, docking the device to
start dreaming, then waiting another minute, will result in the hub
closing due to the timeout.

Bug: 323247392
Fixed: 323247392
Test: atest CommunalSceneStartableTest
      also manually verified on device
Flag: ACONFIG com.android.systemui.communal_hub TEAMFOOD
Change-Id: I3c6639035f579226d334ce96040ec03a631076e0