IMMS: Disallow BAL of INPUT_METHOD_SETTINGS via CLIENT_INTENT
Fixes a vector for apps launching the IME settings from the background.
This isn't exactly useful for malicious apps because the PendingIntent
is immutable, but because mutating immutable intents silently fails,
a malicious app might unintentionally launch the original intent.

So while not exploitable for BAL purposes, this can still be annoying
to users. To avoid that, disallow propagating the creator's BAL
permissions to the PendingIntent.

Fixes: 328014306
Fixes: 332472302
Test: n/a
Flag: NONE
Change-Id: I09ba6070a50b355fc57fe943a20a2da167ff6047