Fix EnterPipToOtherOrientation failure
The expected behavior is to update the pip task bounds on transition
finish. However, after adding the deferWindowLayout, there can be a
racing condition to update the pip task bounds before the transition
starts.

Revert the deferWindowLayout since it is not enough to bundle all the
enter pip ClientTransaction anyway.

Bug: 334006268
Test: EnterPipToOtherOrientation
Change-Id: I6aa3957b61a62c8e690e56291859d817157903e0